Prevent file from being removed if the progress is 100, add additional check to the removeFile method

+++ resources/views/components/uploader.blade.php
@@ -251,6 +251,10 @@
                 },
 
                 removeFile(file) {
+                    if (file.status !== 'uploading') {
+                        return;
+                    }
+                    
                     this.uppy.removeFile(file.id);
 
                     delete this.files[file.id];
